// packages/structurizr-dsl/src/interfaces/ElementType.ts
export enum ElementType {
    Group = "Group",
    Person = "Person",
    SoftwareSystem = "Software System",
    Container = "Container",
    Component = "Component",
    DeploymentNode = "Deployment Node",
    InfrastructureNode = "Infrastructure Node",
    ContainerInstance = "Container Instance",
    SoftwareSystemInstance = "Software System Instance",
}